Abstracted HD variation definitions, after Miryoku
Similar to Miryoku, All HD variation info is now in a separate ##-config.h and ##-adaptive.c files.
All keyboard defs now refer to this abstraction.
All other layers are similarly fully abstracted.
This means a change to a variation or any layer will be reflected in all keyboards.
This is in no way interoperable with Miryoku. It differs greatly, preserving much more granularity to enable combo defs to be based on the abstractions, as well.